refactor!: massive overhaul of cache and broker
The main reason behind this refactor is related to the usage of the
`select!` API: it does not compose well and it is extremely hard to use
in a correct way. Indeed, we were not using it correctly.

I ported domocache and domobroker to a new approach based on the
concepts expressed by the `futures-concurrency` crate. Now, instead of
_select_-ing `Future`s and `Stream`s, we are actually joining, merging
and performing races taking advantage of the type system. This approach
makes clear that channel-like interfaces are needed, and that libp2p is
not directly compatible with this method, requiring some level of
adaptation.

There are still some uses of `select!` left around, hopefully this is
able to fix most of the issues related to async cancellation.

Co-Authored-By: Luca Barbato <lu_zero@gentoo.org>